# resources/openapi-llmstatus.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LLM Stats API",
    "description": "Public API for querying AI model benchmarks, scores, rankings, and pricing.",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/models": {
      "get": {
        "summary": "List Models",
        "operationId": "list_models_v1_models_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "organization",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Organization"
            }
          },
          {
            "name": "family",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Family"
            }
          },
          {
            "name": "modality",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modality"
            }
          },
          {
            "name": "open_weight",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Open Weight"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "min_context",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Context"
            }
          },
          {
            "name": "max_input_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Input Price"
            }
          },
          {
            "name": "max_output_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Output Price"
            }
          },
          {
            "name": "released_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Released After"
            }
          },
          {
            "name": "released_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Released Before"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsListResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models/{model_id}": {
      "get": {
        "summary": "Get Model",
        "operationId": "get_model_v1_models__model_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "param": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param"
          },
          "help_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Help Url"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationRef"
          },
          "family": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FamilyRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "license": {
            "$ref": "#/components/schemas/LicenseRef"
          },
          "open_weight": {
            "type": "boolean",
            "title": "Open Weight"
          },
          "model_type": {
            "type": "string",
            "title": "Model Type"
          },
          "modalities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modalities"
          },
          "context_window": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Window"
          },
          "param_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param Count"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderPricing"
            },
            "type": "array",
            "title": "Providers"
          },
          "top_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Top Scores"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkScore"
            },
            "type": "array",
            "title": "Scores"
          },
          "sources": {
            "$ref": "#/components/schemas/ModelSources"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "organization",
          "license",
          "open_weight",
          "model_type",
          "modalities",
          "providers",
          "top_scores",
          "created_at",
          "updated_at",
          "source",
          "url",
          "scores",
          "sources"
        ],
        "title": "ModelDetail"
      },
...
      "ModelSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationRef"
          },
          "family": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FamilyRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "license": {
            "$ref": "#/components/schemas/LicenseRef"
          },
          "open_weight": {
            "type": "boolean",
            "title": "Open Weight"
          },
          "model_type": {
            "type": "string",
            "title": "Model Type"
          },
          "modalities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modalities"
          },
          "context_window": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Window"
          },
          "param_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param Count"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderPricing"
            },
            "type": "array",
            "title": "Providers"
          },
          "top_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Top Scores"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "organization",
          "license",
          "open_weight",
          "model_type",
          "modalities",
          "providers",
          "top_scores",
          "created_at",
          "updated_at",
          "source",
          "url"
        ],
        "title": "ModelSummary"
      },
      "ModelsListResponse": {
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelSummary"
            },
            "type": "array",
            "title": "Models"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "models",
          "total"
        ],
        "title": "ModelsListResponse"
      },
      "OrganizationRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "OrganizationRef"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "/stats"
    }
  ]
}